+++ core/tauri/scripts/core.js
@@ -17,13 +17,9 @@
 
   window.__TAURI__.convertFileSrc = function convertFileSrc(filePath, protocol = 'asset') {
     const path = encodeURIComponent(filePath)
-    return osName === 'windows'
+    return osName === 'windows' || osName === 'android'
       ? `http://${protocol}.localhost/${path}`
-      : (
-        osName === 'android'
-          ? `https://${protocol}.localhost/${path}`
-          : `${protocol}://localhost/${path}`
-      )
+      : `${protocol}://localhost/${path}`
   }
 
   window.__TAURI__.transformCallback = function transformCallback(