+++ core/tauri/CHANGELOG.md
@@ -1171,7 +1171,7 @@ Here is the logic flow that determines if JSON or JSON5 will be used to parse th
 
 ## \[1.0.0-beta.8]
 
-- Fix missing asset protocol path.Now the protocol is `https://asset.localhost/path/to/file` on Windows. Lunix and macOS
+- Fix missing asset protocol path.Now the protocol is `http://asset.localhost/path/to/file` on Windows. Lunix and macOS
   is still `asset://path/to/file`.
   - [994b5325](https://www.github.com/tauri-apps/tauri/commit/994b5325dd385f564b37fe1530c5d798dc925fff) fix: missing asset protocol path ([#2484](https://www.github.com/tauri-apps/tauri/pull/2484)) on 2021-08-23
 - **Breaking change:** Removed `register_uri_scheme_protocol` from the `WebviewAttributes` struct and renamed `register_global_uri_scheme_protocol` to `register_uri_scheme_protocol` on the `Builder` struct, which now takes a `Fn(&AppHandle, &http::Request) -> http::Response` closure.