+++ core/tauri/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## \[2.0.0-alpha.14]
+
+### Bug Fixes
+
+- [`d5074af5`](https://www.github.com/tauri-apps/tauri/commit/d5074af562b2b5cb6c5711442097c4058af32db6)([#7801](https://www.github.com/tauri-apps/tauri/pull/7801)) Fixes custom protocol not working on Windows.
+
+### Dependencies
+
+- Upgraded to `tauri-runtime-wry@1.0.0-alpha.2`
+
+### Breaking Changes
+
+- [`d5074af5`](https://www.github.com/tauri-apps/tauri/commit/d5074af562b2b5cb6c5711442097c4058af32db6)([#7801](https://www.github.com/tauri-apps/tauri/pull/7801)) The custom protocol on Android now uses the `http` scheme instead of `https`.
+
 ## \[2.0.0-alpha.13]
 
 ### Breaking Changes